Move vikunja operators to concept namespace
